Inline errors into frame_errors
