+++ scylla-cql/src/frame/frame_errors.rs
@@ -1,24 +1,150 @@
 //! Low-level errors that can occur during CQL frame parsing and serialization.
 
 use std::error::Error;
+use std::num::TryFromIntError;
 use std::sync::Arc;
 
 use super::TryFromPrimitiveError;
 use super::response::CqlResponseKind;
+use crate::serialize::SerializationError;
 use crate::utils::parse::ParseErrorCause;
 use thiserror::Error;
 
-pub use super::request::{
-    auth_response::AuthResponseSerializationError,
-    batch::{BatchSerializationError, BatchStatementSerializationError},
-    execute::ExecuteSerializationError,
-    prepare::PrepareSerializationError,
-    query::{QueryParametersSerializationError, QuerySerializationError},
-    register::RegisterSerializationError,
-    startup::StartupSerializationError,
-};
-
 pub use scylla_cql_core::frame::frame_errors::LowLevelDeserializationError;
+
+/// An error type returned when serialization of AUTH_RESPONSE request fails.
+#[non_exhaustive]
+#[derive(Error, Debug, Clone)]
+pub enum AuthResponseSerializationError {
+    /// Maximum response's body length exceeded.
+    #[error("AUTH_RESPONSE body bytes length too big: {0}")]
+    ResponseSerialization(TryFromIntError),
+}
+
+/// An error type returned when serialization of BATCH request fails.
+#[non_exhaustive]
+#[derive(Error, Debug, Clone)]
+pub enum BatchSerializationError {
+    /// Maximum number of batch statements exceeded.
+    #[error(
+        "Too many statements in the batch. Received {0} statements, when u16::MAX is maximum possible value."
+    )]
+    TooManyStatements(usize),
+
+    /// Number of batch statements differs from number of provided bound value lists.
+    #[error(
+        "Number of provided value lists must be equal to number of batch statements (got {n_value_lists} value lists, {n_statements} statements)"
+    )]
+    ValuesAndStatementsLengthMismatch {
+        n_value_lists: usize,
+        n_statements: usize,
+    },
+
+    /// Failed to serialize a statement in the batch.
+    #[error("Failed to serialize batch statement. statement idx: {statement_idx}, error: {error}")]
+    StatementSerialization {
+        statement_idx: usize,
+        error: BatchStatementSerializationError,
+    },
+
+    /// Number of announced batch statements differs from actual number of batch statements.
+    #[error(
+        "Invalid Batch constructed: not as many statements serialized as announced (announced: {n_announced_statements}, serialized: {n_serialized_statements})"
+    )]
+    BadBatchConstructed {
+        n_announced_statements: usize,
+        n_serialized_statements: usize,
+    },
+}
+
+/// An error type returned when serialization of one of the
+/// batch statements fails.
+#[non_exhaustive]
+#[derive(Error, Debug, Clone)]
+pub enum BatchStatementSerializationError {
+    /// Failed to serialize the CQL statement string.
+    #[error("Failed to serialize unprepared statement's content: {0}")]
+    StatementStringSerialization(TryFromIntError),
+
+    /// Maximum value of statement id exceeded.
+    #[error("Malformed prepared statement's id: {0}")]
+    StatementIdSerialization(TryFromIntError),
+
+    /// Failed to serialize statement's bound values.
+    #[error("Failed to serialize statement's values: {0}")]
+    ValuesSerialiation(SerializationError),
+
+    /// Too many bound values provided.
+    #[error("Too many values provided for the statement: {0}")]
+    TooManyValues(usize),
+}
+
+/// An error type returned when serialization of EXECUTE request fails.
+// TODO(2.0): Remove "Serialization" suffix from error names.
+#[expect(clippy::enum_variant_names)]
+#[non_exhaustive]
+#[derive(Error, Debug, Clone)]
+pub enum ExecuteSerializationError {
+    /// Failed to serialize query parameters.
+    #[error("Malformed query parameters: {0}")]
+    QueryParametersSerialization(QueryParametersSerializationError),
+
+    /// Failed to serialize prepared statement id.
+    #[error("Malformed statement id: {0}")]
+    StatementIdSerialization(TryFromIntError),
+
+    #[error("Malformed result metadata id: {0}")]
+    ResultMetadataIdSerialization(TryFromIntError),
+}
+
+/// An error type returned when serialization of PREPARE request fails.
+#[non_exhaustive]
+#[derive(Error, Debug, Clone)]
+pub enum PrepareSerializationError {
+    /// Failed to serialize the CQL statement string.
+    #[error("Failed to serialize statement contents: {0}")]
+    StatementStringSerialization(TryFromIntError),
+}
+
+/// An error type returned when serialization of QUERY request fails.
+#[non_exhaustive]
+#[derive(Error, Debug, Clone)]
+pub enum QuerySerializationError {
+    /// Failed to serialize query parameters.
+    #[error("Invalid query parameters: {0}")]
+    QueryParametersSerialization(QueryParametersSerializationError),
+
+    /// Failed to serialize the CQL statement string.
+    #[error("Failed to serialize a statement content: {0}")]
+    StatementStringSerialization(TryFromIntError),
+}
+
+/// An error type returned when serialization of query parameters fails.
+#[non_exhaustive]
+#[derive(Error, Debug, Clone)]
+pub enum QueryParametersSerializationError {
+    /// Failed to serialize paging state.
+    #[error("Malformed paging state: {0}")]
+    BadPagingState(#[from] TryFromIntError),
+}
+
+/// An error type returned when serialization of REGISTER request fails.
+#[non_exhaustive]
+#[derive(Error, Debug, Clone)]
+pub enum RegisterSerializationError {
+    /// Failed to serialize event types list.
+    #[error("Failed to serialize event types list: {0}")]
+    EventTypesSerialization(TryFromIntError),
+}
+
+/// An error type returned when serialization of STARTUP request fails.
+#[non_exhaustive]
+#[derive(Error, Debug, Clone)]
+pub enum StartupSerializationError {
+    /// Failed to serialize startup options.
+    #[error("Malformed startup options: {0}")]
+    OptionsSerialization(TryFromIntError),
+}
 
 /// An error returned by `parse_response_body_extensions`.
 ///
