+++ scylla-cql/src/frame/request/prepare.rs
@@ -1,15 +1,14 @@
 //! CQL protocol-level representation of a `PREPARE` request.
 
-use std::num::TryFromIntError;
-
-use thiserror::Error;
-
 use crate::frame::frame_errors::CqlRequestSerializationError;
 
 use crate::{
     frame::request::{RequestOpcode, SerializableRequest},
     frame::types,
 };
+
+// Re-export for backward compatibility.
+pub use crate::frame::frame_errors::PrepareSerializationError;
 
 /// CQL protocol-level representation of an `PREPARE` request,
 /// used to prepare a single statement for further execution.
@@ -26,13 +25,4 @@ impl SerializableRequest for Prepare<'_> {
             .map_err(PrepareSerializationError::StatementStringSerialization)?;
         Ok(())
     }
-}
-
-/// An error type returned when serialization of PREPARE request fails.
-#[non_exhaustive]
-#[derive(Error, Debug, Clone)]
-pub enum PrepareSerializationError {
-    /// Failed to serialize the CQL statement string.
-    #[error("Failed to serialize statement contents: {0}")]
-    StatementStringSerialization(TryFromIntError),
 }
