+++ scylla-cql/src/frame/request/register.rs
@@ -1,15 +1,14 @@
 //! CQL protocol-level representation of a `REGISTER` request.
 
-use std::num::TryFromIntError;
-
-use thiserror::Error;
-
 use crate::frame::{
     frame_errors::CqlRequestSerializationError,
     request::{RequestOpcode, SerializableRequest},
     server_event_type::{EventType, EventTypeV2},
     types,
 };
+
+// Re-export for backward compatibility.
+pub use crate::frame::frame_errors::RegisterSerializationError;
 
 /// The CQL protocol-level representation of an `REGISTER` request,
 /// used to subscribe for server events.
@@ -55,13 +54,4 @@ impl SerializableRequest for RegisterV2 {
             .map_err(RegisterSerializationError::EventTypesSerialization)?;
         Ok(())
     }
-}
-
-/// An error type returned when serialization of REGISTER request fails.
-#[non_exhaustive]
-#[derive(Error, Debug, Clone)]
-pub enum RegisterSerializationError {
-    /// Failed to serialize event types list.
-    #[error("Failed to serialize event types list: {0}")]
-    EventTypesSerialization(TryFromIntError),
 }
